# lib/api-zod/src/generated/types/changeType.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API for comparing redactions of legal acts (NPA)
 * OpenAPI spec version: 0.1.0
 */

export type ChangeType = typeof ChangeType[keyof typeof ChangeType];


export const ChangeType = {
  addition: 'addition',
  deletion: 'deletion',
  replacement: 'replacement',
  move: 'move',
} as const;